requirements.txt
alembic==1.18.0
attrs==25.4.0
blinker==1.9.0
boto3==1.42.25
botocore==1.42.25
cachelib==0.13.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==39.0.0
dnspython==2.8.0
eventlet==0.40.4
Flask==2.3.0
Flask-Caching==2.3.1
flask-expects-json==1.7.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5